Fix several more -Wtype-limits warnings (cf. r3759)
Embarrassingly, many seem to be from my 64K sidedefs patch (r3588).

Two remain: one in jsoncpp, which is an embedded upstream library, not
part of Odamex; and another in sv_sqp.cc, which is a comparison against
a version number macro which might not always expand to zero.

Signed-off-by: RjY <rjy@users.sourceforge.net>

SVN r3760 (trunk)