Fix a longstanding (at least 5 year old) display bug in the unused clc_changeteam message.
This message presumably allows you to change your team to something specific.  However, the client has no way to actually send this message to the server, since teams are usually changed in userinfo.  However, in addition to being the source of an out-of-bounds team array access a few commits ago, the team join message didn't even render correctly in the first place due to mismatched tokens in the format string.

SVN r4111 (trunk)