Fix up P_LookForPlayers().
It passes our test suite just as well as the the old function, but is much clearer and doesn't rely on direct access to the players vector.  I also think it covers more corner cases with gaps in playeringame, but I don't think we have a demo that covers that.

SVN r4120 (trunk)