Begin cloning sector thinkers for predicted snaps
When copying a sector for prediction, P_CopySector() previously made a
shallow copy that copied thinker pointers without also making a copy of
the thinkers.  This meant that any thunks done on the prediction sector
had effects on the actual gamesim, resulting in hilarious (and
annoying) bugs.

To remedy this, I am introducing a Clone() method for sector effects
which actually deep-clones the entire thinker without actually adding
it to the thinker list.  Currently only DDoor::Clone() works - other
sector effects will follow.

As a precaution during development, calling the base Clone method will
fatal error the client, which it currently does for anything that isn't
a DDoor.