Remove CTF status bar indicators
The CTF indicators on the status bar took the place of keys and lit up
in strange ways depending on the state of each flag.  The CTF
indicators used in the fullscreen HUD are much easier to make
sense of, so this commit removes the status bar indicators and always
uses the fullscreen CTF indicators.