Do not set snapshot z axis for Actors on ground
Before this commit, standing on a platform that was not locally
predicted had a very good chance of pushing you into the floor or
floating you above the ground due to prediction inaccuracy.  Now, we
simply assume that Actors on the ground already have their proper z
axis set and ignore any attempts to override it from a snapshot.