Fix fencepost error in alpha calculation
translevel appears to be a fixed_t that ranges from 0.0 to 1.0, but the
prior constant was only masking off the lower 10 bits, which meant that
any value exactly 1.0 or greater caused bga to underflow to a negative
number, which later causes a crash.  This change masks off the bits
past 16 as well, so no more fencepost.