Fix fencepost error in transparency calculation
Turns out my original idea was faulty because the decimal was
truncated, leaving only the fractional part, which was wrong-headed.
Instead, we simply take the minimum of our transparency calculation and
255.  This results in a slowdown on the order of nanoseconds in a debug
build and is indistinguishable from the broken implementation in an
optimized one...except this one of course doesn't crash.