Add support for MUSIC_* strings
I think this is a BEX extension that allows you to change which music
lump is used for which level.  Adding support for this required updating
the LANGUAGE lump, so I went ahead and added the base LANGUAGE file that
you're supposed to pass to the converter as well.

BEX support itself is untested, but Back to Saturn X E2's music plays
correctly now with ZMAPINFO.