Introduce StrFormat and VStrFormat
These functions act like asprintf, except instead of formatting to an
allocated char* it allocates to an existing passed std::string.  The
reason I chose this approach was that I wanted to leave open the
possibility of reusing the same std::string multiple times and only
incuring the cost of a single internal allocation.

This commit also includes a sample use elsewhere in the code.