Add flag to force color diagnostics
This is specifically to get colors working with Ninja.  Ninja runs the
compilers in a way that causes them to turn their colors off.  However,
we can't simply force colors on all the time, because it might break
IDE's that use Ninja as a build tool.  This flag is the best compromise.