Attempt to fix crashes related to invalid teams.
The real issue here is the TEAM_NONE that can be called without issue, as it's the default "invalid" team.
As a result, the game freaks out and calls a fatal error.
To fix it, I had to check it when connecting, or switching teams,
and set it back to the default team (AKA BLUE).