Synchronize properly monster total & fix the archvile resurrect sound not being played online.
Similarly to many other Doom Ports, we create a separate counter for respawned monsters. That way, it fixes an issue where a failed survival map will restart with "total+respawned monsters", as the calculation is done when a new map is started, not when it's restarted.
Besides, the client cannot access those parts of the code where monsters respawn. The only way is to ask a server to give the respawned counter, and we do that after each monster respawn/resurrect.
Also changed the "The match has started" line for coop. It's a game, not a match.