Send for each damaged player their attacker netID.
When your player is hit, no matter their direction, the mugface is still looking towards us, even if the attack comes from the left and right.
In fact, the server doesn't send out the attacker, and Odamex still believes the attacker is 0 (so a hazard or self-attack).

This is fixed by sending the netid of the attacker (if any) to each damageplayer packet.

Fixes issue #239.