Fix the infamous door desyncs.
Starting from version 0.6.4 with the rewrite, doors were desyncing, with a rare possibility of being unable to reopening doors despite being truly closed. As a consequence, a few D2ALL demos desynced when trying to open doors.
The problem was because there was a 1-unit too deep check during the movement of the sector.