Fix MAP 8 sound issues.
We remove the fact that the sound is way louder the closer it is to the player. However, we're doing a small hack that only leaves the minimum attenuation, so that it's still a minimum vanilla accurate.

Also remove the louder sounds that could happen too on those maps when using ZDoom sound compatibility.