Moved MF_SPECTATOR to an Odamex Flag.
As a result, I had to also add a new value for the visplane class called mobjflags2, made specifically for Odamex flags.
It's probably not the cleanest solution right now, but for now it does the job just fine.

Also warn the user of the partial or incomplete MBF flags if user has developer 1 set.