Fix occasional crashes caused by OOB fuzzer read
The existing code that turned the fuzz offset into a pitch offset did
not return results that were accurate.  This was causing an OOB read of
the screen buffer when you had a spectre in the lower right corner of
your vision.  I replaced the old bit-twiddle cleverness with a multiply.