Allow building against the system miniupnpc library
`USE_INTERNAL_MINIUPNP` defaults to true and ignores `USE_INTERNAL_LIBS` because
users are unlikely to have it installed.

Although miniupnpc uses pkg-config, it doesn't add any include paths to the
flags because it expects your include directives to include the miniupnpc
directory. We should therefore do the same with the internal build so that
either can be used.

Bug: https://github.com/odamex/odamex/issues/261